feat: 함수 쪼개기, 하늘 컬러 배리에이션
